llm-job/requirements.txt
traceloop-sdk>=0.50.1
openai>=2.13.0

